Refactor Dialog interface and improve functionality
This update brings changes to the Dialog.svelte file in Layout/sidebar, drastically refactoring the interface both visually and functionally. It brings in new imports (like DropdownMenu and Popover), creates checkboxes and adds new functionalities like selecting a label or due date. This change also introduces status options for notes with an ability to navigate the form utilizing the keyboard, increasing its accessibility.